Added support for VR rendering via WebVR
The distortion is now handled by the navigator.
We only retrieve the field of view and ideal render target resolution for the current hmd and setup a double camera node.
The dualcam has been updated, you can press f on a supported navigator with an oculus and it just works.