Fixes error in world matrix computation
The viewer pushed two matrix in modelview to initialize the culling. The
first was an matrix identity to validate the call to
addPositionedAttribute for lights setup, and the second was the
camera. The cull visitor was expecting the first matrix at index 0 to be
the camera. So the the first matrix computed did not used the camera
one. Instead of shifting the index of the tracking of the camera in cull
visitor, I fixed the lights setup. There is no need to push a identity
matrix as first for the the lights. I just changed the function
signature to use the same used in osg, meaing addPositionnedAttribute (
matrix, attribute ).
The side effect is that if you called yourself addPositionnedAttribute,
you will have to update to the new signature. I did not make backward
compatibilty because I dont think a lot of people call it explicitely.