Adds in OrbitManipulator computeHomePosition( useBoundingBox )
To center the view only on visible objects in the scene, It needed to use the
ComputeBoundsVisitor.
Using ComputeBoundsVisitor on the node used by the OrbitManipulator, the
manipulator get the real bounding box.
A side effect is that when using this method, the Absolute camera are
not skipped, so this commit also add setBound in Geometry to specify an
invalid boundox typically for skydom.
To make the method more consistent every expand* method in bounding box
and bounding sphere have now an explicit name.