Simplify FBO api
Note that we could almost remove attachement code in fbo because it's
used in RenderStage and if you need to custmize it, you will probably
build the fbo by its api.
Another point is that we keep gl context to manipulate the FBO /
Texture. That would be better to keep instead a reference to State like
it's done in osg.