[#DEV-6488] Robustifies osg.image.isReady
Fixes previous osg Texture code which was using a very relaxed test to check if it has an HTML image or a OSG image inside, which gives incorrect results to user with ublock extension.

Ublock Extension add a HTML image "isReady" method, with exaclty the same name as osg Image leading to current code believing incorrectly that the "wrapped by ublock Image" is an osg Image, or is not a HTML image thus never loading it in a gpu texture.

Any inconsiderate browser extension wrapping the Image html element,
and there may be even more in the future when browser will freeze the
'natives' objects like Image, thus needs a "detection by feature" as
last resort of the isReady method.