Removes extra pushStateSet when using insertStateSet
When using insertStateSet and RenderLeaf use this codepath

```
var hash = state.getStateSetStackHash();
if ( previousHash !== hash ) {
    this._parent.moveStateGraph( state, undefined, this._parent.parent );
    state.applyStateSet( this._parent.stateset );
    previousHash = hash;
}
```

We where pushing all StateSet from the stateGraph ( from current to root
) on State.
So for example we had something like that in state after rendering
multiple RenderLeaf
[ 25, 39, 46, 567, 248, 251, 569, 565, 55, 257, 39, 46, 567, 248, 251,
569, 565, 55, 257, ... ]
the sequence [ 39, 46, 567, 248, 251, 569, 565, 55, 257 ] was repeated
because of moveStateGraph( state, undefined, this._parent.parent );
Actually this code path needs to re apply State's StateSet if an
insertStateSet has been used but does not need to push/pop any StateSet.
To re apply StateSet we still need to apply the one from the current
StateGraph because it is not pushed on the State's stack. So the fix is
just to remove the line:
```
moveStateGraph( state, undefined,this._parent.parent );
```