[#D3D-3238] Fixes morph animation linking
Animation linking is made using a uniqueTargetName: targetName + '.' + channelName
For morph, it was built with targetName + '.' + morphTargetIndex, that worked until now
since channelName == targetIndex. For unicity reasons, channelName can be different
so the linking doesn't work.
Remplacing the index by the actual channelName (that corresponds to the name of the updateMorph element here).