Adds node testing, replaces phantom.js
Fixes hang in tests in nodejs context

It was due to the method initEvents in viewer. Because in nodejs we dont really
want to handle this I did a mockup on Viewer class that call initEvents only
in browser context. It means there is one test less in nodejs context related
to device init.

Fixes benchmarks and gruntfile